arch/risc-v/qemu-rv: Support both rv32/rv64 core
