+++ arch/risc-v/Kconfig
@@ -110,9 +110,8 @@ config ARCH_CHIP_RV32M1
 	---help---
 		NXP RV32M1 processor (RISC-V Core with PULP extensions).
 
-config ARCH_CHIP_QEMU_RV32
-	bool "QEMU RV32"
-	select ARCH_RV32
+config ARCH_CHIP_QEMU_RV
+	bool "QEMU RV"
 	select ARCH_RV_ISA_F
 	select ARCH_RV_ISA_D
 	---help---
@@ -176,7 +175,7 @@ config ARCH_CHIP
 	default "c906"		if ARCH_CHIP_C906
 	default "mpfs"		if ARCH_CHIP_MPFS
 	default "rv32m1"	if ARCH_CHIP_RV32M1
-	default "qemu-rv32"	if ARCH_CHIP_QEMU_RV32
+	default "qemu-rv"	if ARCH_CHIP_QEMU_RV
 
 config ARCH_RISCV_INTXCPT_EXTENSIONS
 	bool "RISC-V Integer Context Extensions"
@@ -221,7 +220,7 @@ endif
 if ARCH_CHIP_RV32M1
 source "arch/risc-v/src/rv32m1/Kconfig"
 endif
-if ARCH_CHIP_QEMU_RV32
-source "arch/risc-v/src/qemu-rv32/Kconfig"
+if ARCH_CHIP_QEMU_RV
+source "arch/risc-v/src/qemu-rv/Kconfig"
 endif
 endif
