--- arch/risc-v/src/qemu-rv32/Kconfig
@@ -1,24 +1,0 @@
-#
-# For a description of the syntax of this configuration file,
-# see the file kconfig-language.txt in the NuttX tools repository.
-#
-
-if ARCH_CHIP_QEMU_RV32
-comment "QEMU RV32 Options"
-
-config ARCH_CHIP_QEMU_RV32_ISA_M
-    bool "Standard Extension for Integer Multiplication and Division"
-    default y
-    select ARCH_RV_ISA_M
-
-config ARCH_CHIP_QEMU_RV32_ISA_A
-    bool "Standard Extension for Atomic Instructions"
-    default y
-    select ARCH_RV_ISA_A
-
-config ARCH_CHIP_QEMU_RV32_ISA_C
-    bool "Standard Extension for Compressed Instructions"
-    default y
-    select ARCH_RV_ISA_C
-
-endif
