--- arch/risc-v/src/qemu-rv32/qemu_rv32_head.S
@@ -1,65 +1,0 @@
-/****************************************************************************
- * arch/risc-v/src/qemu-rv32/qemu_rv32_head.S
- *
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.  The
- * ASF licenses this file to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance with the
- * License.  You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
- * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.  See the
- * License for the specific language governing permissions and limitations
- * under the License.
- *
- ****************************************************************************/
-
-/****************************************************************************
- * Included Files
- ****************************************************************************/
-
-#include <nuttx/config.h>
-#include <arch/irq.h>
-
-#include "chip.h"
-
-/****************************************************************************
- * Public Symbols
- ****************************************************************************/
-
-  /* Exported Symbols */
-
-  .section .text
-  .global __start
-
-__start:
-
-  /* Set stack pointer to the idle thread stack */
-
-  lui  sp, %hi(QEMU_RV32_IDLESTACK_TOP)
-  addi sp, sp, %lo(QEMU_RV32_IDLESTACK_TOP)
-
-	csrw	mie, zero
-
-  la   t0, __trap_vec
-  csrw mtvec, t0
-
-  /* Jump to qemu_rv32_start */
-
-  jal  x1, qemu_rv32_start
-
-  /* We shouldn't return from _start */
-
-  .global _init
-  .global _fini
-
-_init:
-_fini:
-
-  /* These don't have to do anything since we use init_array/fini_array. */
-
-  ret
