--- arch/risc-v/src/qemu-rv32/qemu_rv32_start.c
@@ -1,101 +1,0 @@
-/****************************************************************************
- * arch/risc-v/src/qemu-rv32/qemu_rv32_start.c
- *
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.  The
- * ASF licenses this file to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance with the
- * License.  You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
- * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.  See the
- * License for the specific language governing permissions and limitations
- * under the License.
- *
- ****************************************************************************/
-
-/****************************************************************************
- * Included Files
- ****************************************************************************/
-
-#include <nuttx/config.h>
-
-#include <nuttx/init.h>
-#include <nuttx/arch.h>
-#include <arch/board/board.h>
-
-#include "riscv_internal.h"
-
-#include "chip.h"
-
-/****************************************************************************
- * Pre-processor Definitions
- ****************************************************************************/
-
-#ifdef CONFIG_DEBUG_FEATURES
-#define showprogress(c) up_putc(c)
-#else
-#define showprogress(c)
-#endif
-
-/****************************************************************************
- * Public Data
- ****************************************************************************/
-
-/* NOTE: g_idle_topstack needs to point the top of the idle stack
- * for CPU0 and this value is used in up_initial_state()
- */
-
-uintptr_t g_idle_topstack = QEMU_RV32_IDLESTACK_TOP;
-
-/****************************************************************************
- * Public Functions
- ****************************************************************************/
-
-/****************************************************************************
- * Name: qemu_rv32_start
- ****************************************************************************/
-
-void qemu_rv32_start(void)
-{
-  uint32_t *dest;
-
-  /* Clear .bss.  We'll do this inline (vs. calling memset) just to be
-   * certain that there are no issues with the state of global variables.
-   */
-
-  for (dest = &_sbss; dest < &_ebss; )
-    {
-      *dest++ = 0;
-    }
-
-  showprogress('A');
-
-#ifdef USE_EARLYSERIALINIT
-  up_earlyserialinit();
-#endif
-
-  showprogress('B');
-
-  /* Do board initialization */
-
-  showprogress('C');
-
-  /* Call nx_start() */
-
-  nx_start();
-
-  while (true)
-    {
-      asm("WFI");
-    }
-}
-
-void riscv_serialinit(void)
-{
-  up_serialinit();
-}
