--- boards/risc-v/qemu-rv32/rv32-virt/configs/nsh/defconfig
@@ -1,65 +1,0 @@
-#
-# This file is autogenerated: PLEASE DO NOT EDIT IT.
-#
-# You can use "make menuconfig" to make any modifications to the installed .config file.
-# You can then do "make savedefconfig" to generate a new defconfig file that includes your
-# modifications.
-#
-# CONFIG_DISABLE_OS_API is not set
-# CONFIG_DISABLE_PSEUDOFS_OPERATIONS is not set
-# CONFIG_NSH_DISABLEBG is not set
-# CONFIG_NSH_DISABLE_CAT is not set
-# CONFIG_NSH_DISABLE_CD is not set
-# CONFIG_NSH_DISABLE_FREE is not set
-# CONFIG_NSH_DISABLE_HELP is not set
-# CONFIG_NSH_DISABLE_LOSMART is not set
-# CONFIG_NSH_DISABLE_LS is not set
-# CONFIG_NSH_DISABLE_MOUNT is not set
-# CONFIG_NSH_DISABLE_PS is not set
-# CONFIG_NSH_DISABLE_PSSTACKUSAGE is not set
-# CONFIG_NSH_DISABLE_UNAME is not set
-CONFIG_16550_ADDRWIDTH=32
-CONFIG_16550_UART0=y
-CONFIG_16550_UART0_BASE=0x10000000
-CONFIG_16550_UART0_CLOCK=3686400
-CONFIG_16550_UART0_IRQ=37
-CONFIG_16550_UART0_SERIAL_CONSOLE=y
-CONFIG_16550_UART=y
-CONFIG_ARCH="risc-v"
-CONFIG_ARCH_BOARD="rv32-virt"
-CONFIG_ARCH_BOARD_QEMU_RV32_VIRT=y
-CONFIG_ARCH_CHIP="qemu-rv32"
-CONFIG_ARCH_CHIP_QEMU_RV32=y
-CONFIG_ARCH_INTERRUPTSTACK=2048
-CONFIG_ARCH_RISCV=y
-CONFIG_ARCH_STACKDUMP=y
-CONFIG_BINFMT_DISABLE=y
-CONFIG_BOARD_LOOPSPERMSEC=6366
-CONFIG_BUILTIN=y
-CONFIG_CLOCK_MONOTONIC=y
-CONFIG_DEBUG_FULLOPT=y
-CONFIG_DEBUG_SYMBOLS=y
-CONFIG_DEFAULT_SMALL=y
-CONFIG_DEV_ZERO=y
-CONFIG_FS_PROCFS=y
-CONFIG_IDLETHREAD_STACKSIZE=2048
-CONFIG_INIT_ENTRYPOINT="nsh_main"
-CONFIG_INTELHEX_BINARY=y
-CONFIG_LIBC_FLOATINGPOINT=y
-CONFIG_LIBC_PERROR_STDOUT=y
-CONFIG_LIBC_STRERROR=y
-CONFIG_NFILE_DESCRIPTORS_PER_BLOCK=6
-CONFIG_NSH_ARCHINIT=y
-CONFIG_NSH_BUILTIN_APPS=y
-CONFIG_NSH_FILEIOSIZE=64
-CONFIG_PREALLOC_TIMERS=0
-CONFIG_RAM_SIZE=33554432
-CONFIG_RAM_START=0x80000000
-CONFIG_RR_INTERVAL=200
-CONFIG_SCHED_WAITPID=y
-CONFIG_SERIAL_UART_ARCH_MMIO=y
-CONFIG_STACK_COLORATION=y
-CONFIG_START_MONTH=12
-CONFIG_START_YEAR=2021
-CONFIG_SYSTEM_NSH=y
-CONFIG_USEC_PER_TICK=1000
