--- boards/risc-v/qemu-rv32/rv32-virt/src/Makefile
@@ -1,27 +1,0 @@
-############################################################################
-# boards/risc-v/rv32m1/rv32m1-vega/src/Makefile
-#
-# Licensed to the Apache Software Foundation (ASF) under one or more
-# contributor license agreements.  See the NOTICE file distributed with
-# this work for additional information regarding copyright ownership.  The
-# ASF licenses this file to you under the Apache License, Version 2.0 (the
-# "License"); you may not use this file except in compliance with the
-# License.  You may obtain a copy of the License at
-#
-#   http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
-# WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.  See the
-# License for the specific language governing permissions and limitations
-# under the License.
-#
-############################################################################
-
-include $(TOPDIR)/Make.defs
-
-RCSRCS = etc/init.d/rcS
-
-CSRCS = qemu_rv32_appinit.c
-
-include $(TOPDIR)/boards/Board.mk
